# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strong Hire API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/{session_id}/debrief": {
      "get": {
        "tags": [
          "scoring"
        ],
        "summary": "Get Debrief",
        "description": "FB-1, FB-2, PR-2. Poll while status is \"scoring\" (FB-3: ready within 60 seconds).",
        "operationId": "get_debrief_sessions__session_id__debrief_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Debrief"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job-targets/{job_target_id}/progress": {
      "get": {
        "tags": [
          "scoring"
        ],
        "summary": "Get Progress",
        "description": "PR-1 (Realistic sessions only) and PR-2.",
        "operationId": "get_progress_job_targets__job_target_id__progress_get",
        "parameters": [
          {
            "name": "job_target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Target Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobProgress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/scoring": {
      "post": {
        "tags": [
          "scoring"
        ],
        "summary": "Score Session",
        "description": "Start scoring an ended session, or retry after a failure.",
        "operationId": "score_session_sessions__session_id__scoring_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Competency": {
        "type": "string",
        "enum": [
          "ownership",
          "impact",
          "collaboration",
          "conflict_handling",
          "learning_from_failure",
          "communication",
          "role_fit",
          "depth_of_experience",
          "judgment",
          "motivation",
          "team_fit",
          "scope_at_level",
          "technical_depth",
          "accuracy",
          "reasoning",
          "trade_offs",
          "clarity_of_explanation",
          "problem_framing",
          "structure",
          "user_and_business_sense",
          "estimation_logic",
          "prioritization",
          "recommendation"
        ],
        "title": "Competency",
        "description": "Scored competencies from the spec, grouped by interview type in COMPETENCIES_BY_TYPE."
      },
...
      "CompetencyScore": {
        "properties": {
          "competency": {
            "$ref": "#/components/schemas/Competency"
          },
          "score": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Score"
          },
          "justification": {
            "type": "string",
            "minLength": 1,
            "title": "Justification"
          },
          "quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Quotes",
            "description": "Transcript quotes backing the score."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "competency",
          "score",
          "justification",
          "quotes"
        ],
        "title": "CompetencyScore"
      },
      "CompetencyTrend": {
        "properties": {
          "competency": {
            "$ref": "#/components/schemas/Competency"
          },
          "sessions": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sessions"
          },
          "first": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "First"
          },
          "latest": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Latest"
          },
          "change": {
            "type": "number",
            "title": "Change",
            "description": "latest minus first"
          },
          "average": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Average"
          },
          "direction": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "flat",
              "single"
            ],
            "title": "Direction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "competency",
          "sessions",
          "first",
          "latest",
          "change",
          "average",
          "direction"
        ],
        "title": "CompetencyTrend",
        "description": "PR-1: one competency over the Realistic sessions of a job target."
      },
...
      "Debrief": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/DebriefSession"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_ended",
              "not_started",
              "scoring",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "scorecard": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scorecard"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannedSession"
              },
              {
                "type": "null"
              }
            ],
            "description": "PR-2. Null until the debrief is ready."
          },
          "generic_mode": {
            "type": "boolean",
            "title": "Generic Mode",
            "description": "True when no company profile was used: no values."
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "values_framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values Framework",
            "description": "Name of the company's values framework, for example 'Leadership Principles'."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session",
          "status",
          "scorecard",
          "next_session",
          "generic_mode",
          "company_name",
          "values_framework"
        ],
        "title": "Debrief"
      },
      "DebriefSession": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Target Id"
          },
          "config": {
            "$ref": "#/components/schemas/SessionConfig"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "minutes_billed": {
            "type": "integer",
            "title": "Minutes Billed"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason",
            "description": "Plain reason when the session failed before it started."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "job_target_id",
          "config",
          "status",
          "started_at",
          "ended_at",
          "minutes_billed"
        ],
        "title": "DebriefSession",
        "description": "The session a debrief belongs to. Same shape as the web app's SessionRecord."
      },
...
      "Difficulty": {
        "type": "string",
        "enum": [
          "friendly",
          "realistic",
          "tough"
        ],
        "title": "Difficulty",
        "description": "IV-4."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HireSignal": {
        "type": "string",
        "enum": [
          "Strong Hire",
          "Hire",
          "Lean Hire",
          "Lean No Hire",
          "No Hire"
        ],
        "title": "HireSignal",
        "description": "FB-1. Ordered from strongest to weakest."
      },
      "InterviewType": {
        "type": "string",
        "enum": [
          "behavioral",
          "hiring_manager",
          "technical_qa",
          "case"
        ],
        "title": "InterviewType",
        "description": "The four v1 interview types (IV-2)."
      },
...
      "JobProgress": {
        "properties": {
          "job_target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Target Id"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/ProgressSnapshot"
            },
            "type": "array",
            "title": "Snapshots",
            "description": "Realistic sessions only (PR-1)."
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/CompetencyTrend"
            },
            "type": "array",
            "title": "Trends",
            "description": "Weakest latest score first."
          },
          "next_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannedSession"
              },
              {
                "type": "null"
              }
            ],
            "description": "PR-2."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_target_id",
          "snapshots",
          "trends",
          "next_session"
        ],
        "title": "JobProgress"
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "new_grad",
          "mid",
          "senior",
          "staff_principal"
        ],
        "title": "Level",
        "description": "Normalized seniority (IV-6). Company-specific labels such as L5 or E5 map to these."
      },
...
      "Mode": {
        "type": "string",
        "enum": [
          "coach",
          "realistic"
        ],
        "title": "Mode",
        "description": "IV-8. Only Realistic sessions count toward progress trends (PR-1)."
      },
...
      "PlannedSession": {
        "properties": {
          "priority": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Priority",
            "description": "1 is the first session to run."
          },
          "interview_type": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty",
            "default": "realistic"
          },
          "focus_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Focus Topics"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "priority",
          "interview_type",
          "focus_topics",
          "reason"
        ],
        "title": "PlannedSession",
        "description": "One recommended practice session (GA-3)."
      },
...
      "ProgressSnapshot": {
        "properties": {
          "job_target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Target Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "competency": {
            "$ref": "#/components/schemas/Competency"
          },
          "score": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Score"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_target_id",
          "session_id",
          "competency",
          "score",
          "at"
        ],
        "title": "ProgressSnapshot",
        "description": "One competency score from one Realistic session, for trend charts (PR-1)."
      },
...
      "QuestionScore": {
        "properties": {
          "question_ref": {
            "type": "string",
            "minLength": 1,
            "title": "Question Ref"
          },
          "question_text": {
            "type": "string",
            "minLength": 1,
            "title": "Question Text"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/CompetencyScore"
            },
            "type": "array",
            "minItems": 1,
            "title": "Scores"
          },
          "value_scores": {
            "items": {
              "$ref": "#/components/schemas/ValueScore"
            },
            "type": "array",
            "title": "Value Scores"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "misses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Misses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_ref",
          "question_text",
          "scores"
        ],
        "title": "QuestionScore"
      },
...
      "Scorecard": {
        "properties": {
          "hire_signal": {
            "$ref": "#/components/schemas/HireSignal"
          },
          "rationale": {
            "type": "string",
            "minLength": 1,
            "title": "Rationale",
            "description": "3 to 5 sentences, like a debrief summary."
          },
          "competency_scores": {
            "items": {
              "$ref": "#/components/schemas/CompetencyScore"
            },
            "type": "array",
            "minItems": 1,
            "title": "Competency Scores",
            "description": "Overall score per competency across the session."
          },
          "value_scores": {
            "items": {
              "$ref": "#/components/schemas/ValueScore"
            },
            "type": "array",
            "title": "Value Scores",
            "description": "Overall score per company value. Empty in generic mode."
          },
          "per_question": {
            "items": {
              "$ref": "#/components/schemas/QuestionScore"
            },
            "type": "array",
            "title": "Per Question"
          },
          "scorer_model": {
            "type": "string",
            "minLength": 1,
            "title": "Scorer Model",
            "description": "Gateway model alias, never a vendor id."
          },
          "rubric_version": {
            "type": "string",
            "minLength": 1,
            "title": "Rubric Version",
            "description": "Prompt ref, for example 'scorer/rubric.v1'."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hire_signal",
          "rationale",
          "competency_scores",
          "scorer_model",
          "rubric_version"
        ],
        "title": "Scorecard"
      },
      "ScoringAccepted": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "job_id",
          "status"
        ],
        "title": "ScoringAccepted"
      },
...
      "SessionConfig": {
        "properties": {
          "interview_type": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "duration_min": {
            "type": "integer",
            "enum": [
              10,
              30,
              45
            ],
            "title": "Duration Min",
            "description": "10 is a mini interview."
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "interview_type",
          "difficulty",
          "mode",
          "duration_min",
          "level"
        ],
        "title": "SessionConfig",
        "description": "What the candidate picks before a session (journey 2, step 1)."
      },
...
      "SessionStatus": {
        "type": "string",
        "enum": [
          "created",
          "in_progress",
          "interrupted",
          "scoring",
          "completed",
          "failed"
        ],
        "title": "SessionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValueScore": {
        "properties": {
          "value": {
            "type": "string",
            "minLength": 1,
            "title": "Value"
          },
          "score": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Score"
          },
          "justification": {
            "type": "string",
            "minLength": 1,
            "title": "Justification"
          },
          "quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Quotes",
            "description": "Transcript quotes backing the score."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value",
          "score",
          "justification",
          "quotes"
        ],
        "title": "ValueScore",
        "description": "Score for one company value (Principle.name), same 1 to 4 rubric as competencies."
      },
...
    }
  }
}